+++ example/lib/main.dart
@@ -35,21 +35,21 @@ class FlutterReduxApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      theme: ThemeData.dark(),
+      theme: new ThemeData.dark(),
       title: title,
-      home: StoreProvider<int>(
+      home: new StoreProvider<int>(
         // Pass the store to the StoreProvider. Any ancestor `StoreConnector`
         // Widgets will find and use this value as the `Store`.
         store: store,
-        child: Scaffold(
-          appBar: AppBar(
-            title: Text(title),
+        child: new Scaffold(
+          appBar: new AppBar(
+            title: new Text(title),
           ),
-          body: Center(
-            child: Column(
+          body: new Center(
+            child: new Column(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                Text(
+                new Text(
                   'You have pushed the button this many times:',
                 ),
                 // Connect the Store to a Text Widget that renders the current
@@ -67,10 +67,12 @@ class FlutterReduxApp extends StatelessWidget {
                 // count. No need to manually manage subscriptions or Streams!
                 new StoreConnector<int, String>(
                   converter: (store) => store.state.toString(),
-                  builder: (context, count) => Text(
-                        count,
-                        style: Theme.of(context).textTheme.display1,
-                      ),
+                  builder: (context, count) {
+                    return new Text(
+                      count,
+                      style: Theme.of(context).textTheme.display1,
+                    );
+                  },
                 )
               ],
             ),
@@ -86,12 +88,14 @@ class FlutterReduxApp extends StatelessWidget {
               // with no parameters. It only dispatches an Increment action.
               return () => store.dispatch(Actions.Increment);
             },
-            builder: (context, callback) => FloatingActionButton(
-                  // Attach the `callback` to the `onPressed` attribute
-                  onPressed: callback,
-                  tooltip: 'Increment',
-                  child: Icon(Icons.add),
-                ),
+            builder: (context, callback) {
+              return new FloatingActionButton(
+                // Attach the `callback` to the `onPressed` attribute
+                onPressed: callback,
+                tooltip: 'Increment',
+                child: new Icon(Icons.add),
+              );
+            },
           ),
         ),
       ),