+++ lib/flutter_redux.dart
@@ -16,8 +16,7 @@ class StoreProvider<S> extends InheritedWidget {
     Key key,
     @required Store<S> store,
     @required Widget child,
-  })
-      : assert(store != null),
+  })  : assert(store != null),
         assert(child != null),
         _store = store,
         super(key: key, child: child);
@@ -122,7 +121,7 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
   ///
   /// This can be useful for dispatching actions that fetch data for your Widget
   /// when it is first displayed.
-  final OnInitCallback onInit;
+  final OnInitCallback<S> onInit;
 
   /// A function that will be run when the StoreConnector is removed from the
   /// Widget Tree.
@@ -131,7 +130,7 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
   ///
   /// This can be useful for dispatching actions that remove stale data from
   /// your State tree.
-  final OnDisposeCallback onDispose;
+  final OnDisposeCallback<S> onDispose;
 
   /// Determines whether the Widget should be rebuilt when the Store emits an
   /// onChange event.
@@ -172,8 +171,7 @@ class StoreConnector<S, ViewModel> extends StatelessWidget {
     this.rebuildOnChange = true,
     this.ignoreChange,
     this.onWillChange,
-  })
-      : assert(builder != null),
+  })  : assert(builder != null),
         assert(converter != null),
         super(key: key);
 
@@ -214,7 +212,7 @@ class StoreBuilder<S> extends StatelessWidget {
   ///
   /// This can be useful for dispatching actions that fetch data for your Widget
   /// when it is first displayed.
-  final OnInitCallback onInit;
+  final OnInitCallback<S> onInit;
 
   /// A function that will be run when the StoreBuilder is removed from the
   /// Widget Tree.
@@ -223,7 +221,13 @@ class StoreBuilder<S> extends StatelessWidget {
   ///
   /// This can be useful for dispatching actions that remove stale data from
   /// your State tree.
-  final OnDisposeCallback onDispose;
+  final OnDisposeCallback<S> onDispose;
+
+  /// A function that will be run on State change.
+  ///
+  /// This can be useful for imperative calls to things like Navigator,
+  /// TabController, etc
+  final OnWillChangeCallback<Store<S>> onWillChange;
 
   /// A function that will be run on State change.
   ///
@@ -238,8 +242,7 @@ class StoreBuilder<S> extends StatelessWidget {
     this.onDispose,
     this.rebuildOnChange = true,
     this.onWillChange,
-  })
-      : assert(builder != null),
+  })  : assert(builder != null),
         super(key: key);
 
   @override
@@ -250,7 +253,7 @@ class StoreBuilder<S> extends StatelessWidget {
       rebuildOnChange: rebuildOnChange,
       onInit: onInit,
       onDispose: onDispose,
-      onWillChange: this.onWillChange,
+      onWillChange: onWillChange,
     );
   }
 }
@@ -262,8 +265,8 @@ class _StoreStreamListener<S, ViewModel> extends StatefulWidget {
   final Store<S> store;
   final bool rebuildOnChange;
   final bool distinct;
-  final OnInitCallback onInit;
-  final OnDisposeCallback onDispose;
+  final OnInitCallback<S> onInit;
+  final OnDisposeCallback<S> onDispose;
   final IgnoreChangeTest<S> ignoreChange;
   final OnWillChangeCallback<ViewModel> onWillChange;
 
@@ -278,8 +281,7 @@ class _StoreStreamListener<S, ViewModel> extends StatefulWidget {
     this.rebuildOnChange = true,
     this.ignoreChange,
     this.onWillChange,
-  })
-      : super(key: key);
+  }) : super(key: key);
 
   @override
   State<StatefulWidget> createState() {